refactor: 코어 웹 바이탈 성능 최적화(#413)
